refactor(frontend_tkinter): Move classes to separated files to simplify tests
