+++ ardupilot_methodic_configurator/frontend_tkinter_flightcontroller_info.py
@@ -17,7 +17,8 @@ from tkinter import ttk
 from ardupilot_methodic_configurator import _, __version__
 from ardupilot_methodic_configurator.annotate_params import Par
 from ardupilot_methodic_configurator.backend_flightcontroller import FlightController
-from ardupilot_methodic_configurator.frontend_tkinter_base import BaseWindow, ProgressWindow
+from ardupilot_methodic_configurator.frontend_tkinter_base import BaseWindow
+from ardupilot_methodic_configurator.frontend_tkinter_progress_window import ProgressWindow
 
 
 class FlightControllerInfoWindow(BaseWindow):
